+++ components/ChatContent/reactMarkdown.tsx
@@ -1,4 +1,5 @@
-import React, { memo } from "react";
+import * as React from "react";
 import ReactMarkdown, { Options } from "react-markdown";
 
-export const MemoizedReactMarkdown: React.FC<Options> = memo(ReactMarkdown);
+export const MemoizedReactMarkdown: React.FC<Options> =
+  React.memo(ReactMarkdown);