+++ components/Drawer/types.ts
@@ -1,5 +1,6 @@
 export interface IDrawerPropTypes extends React.PropsWithChildren {
   className?: string;
+  title?: React.ReactNode;
   onClose?: () => void;
   open?: boolean;
   width?: number | string;