+++ components/Menu/Mobile/index.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
 import classNames from "classnames";
+import { useTranslation } from "next-i18next";
 import { useDateFormat } from "l-hooks";
 import { AiOutlineDelete } from "react-icons/ai";
 import { BsChatSquareText } from "react-icons/bs";
@@ -9,6 +10,7 @@ import { useChannel, initChannelList } from "@/hooks";
 import { useMobileMenuOpen } from "@/state";
 
 const MobileMenu: React.FC = () => {
+  const { t } = useTranslation("menu");
   const [channel, setChannel] = useChannel();
   const open = useMobileMenuOpen((state) => state.open);
   const setOpen = useMobileMenuOpen((state) => state.update);
@@ -41,7 +43,8 @@ const MobileMenu: React.FC = () => {
 
   const onDeleteChannel = (e: any, id: string) => {
     e.stopPropagation();
-    if (!confirm("Are you sure delete this conversation?")) return;
+    const confirmMsg = t("delete-conversation");
+    if (!confirm(confirmMsg)) return;
 
     if (channel.list.length <= 1) {
       setChannel((channel) => {
@@ -61,15 +64,21 @@ const MobileMenu: React.FC = () => {
   };
 
   return (
-    <Drawer className="md:hidden" width="85%" open={open} onClose={onClose}>
+    <Drawer
+      className="md:hidden"
+      title={t("coversation-list")}
+      width="85%"
+      open={open}
+      onClose={onClose}
+    >
       <div className="p-2">
         <div
           onClick={onAddChannel}
           className="rounded-lg cursor-pointer flex bg-[#678fff] h-12 text-white mb-2 transition-colors justify-center items-center hover:bg-[#678fff]/80"
         >
-          New Chat
+          {t("new-chat")}
         </div>
-        <div className="h-mobileMenu pr-2 overflow-y-auto">
+        <div className="h-mobileMenu overflow-y-auto">
           {channel.list.map((item) => (
             <div
               key={item.channel_id}
@@ -82,7 +91,7 @@ const MobileMenu: React.FC = () => {
               <div className="flex justify-between items-center">
                 <div className="font-medium text-sm text-ellipsis max-w-[22ch] pl-5 overflow-hidden whitespace-nowrap relative">
                   <BsChatSquareText className="top-[50%] left-0 translate-y-[-50%] absolute" />
-                  {item.channel_name || "New Conversation"}
+                  {item.channel_name || t("new-conversation")}
                 </div>
                 <div className="transition-opacity text-[#858b96] group-hover:opacity-0">
                   {item.chat_list.length
@@ -96,7 +105,7 @@ const MobileMenu: React.FC = () => {
                 </div>
               </div>
               <div className="text-[#858b96] group-hover:text-[#6e737b]">
-                {item.chat_list.length} messages
+                {item.chat_list.length} {t("messages")}
               </div>
               <AiOutlineDelete
                 size={20}