+++ components/Navbar/index.tsx
@@ -1,14 +1,19 @@
-import React from "react";
+import * as React from "react";
+import { useTranslation } from "next-i18next";
 import { useMobileMenuOpen } from "@/state";
 import { Modal } from "@/components";
 import { AiOutlineMenuUnfold, AiOutlineSetting } from "react-icons/ai";
 import { useOpenAIKey, useChannel } from "@/hooks";
 
 const Navbar: React.FC = () => {
+  const { t: tMenu } = useTranslation("menu");
+  const { t: tNav } = useTranslation("nav");
   const [openAIKey, setOpenAIKey] = useOpenAIKey();
   const [channel] = useChannel();
   const [open, setOpen] = React.useState(false);
   const setMobileMenuOpen = useMobileMenuOpen((state) => state.update);
+
+  const settingText = tNav("setting");
 
   const onOpenMenu = () => setMobileMenuOpen(true);
 
@@ -33,8 +38,8 @@ const Navbar: React.FC = () => {
         </div>
         <div className="text-ellipsis max-w-[50%] whitespace-nowrap overflow-hidden">
           {openAIKey
-            ? activeChannel?.channel_name || "New Conversation"
-            : "Set Your OpenAI Key"}
+            ? activeChannel?.channel_name || tMenu("new-conversation")
+            : tNav("set-openai-key")}
         </div>
         <div
           onClick={onOpenSetting}
@@ -43,7 +48,7 @@ const Navbar: React.FC = () => {
           <AiOutlineSetting size={24} />
         </div>
       </div>
-      <Modal title="Setting" width={600} open={open} onClose={onClose}>
+      <Modal title={settingText} width={600} open={open} onClose={onClose}>
         <div className="border-b flex border-slate-100 py-2 px-6 items-center justify-between">
           <div className="font-semibold text-sm">API key</div>
           <div className="text-xs">
@@ -57,7 +62,7 @@ const Navbar: React.FC = () => {
           </div>
         </div>
         <div className="border-b flex border-slate-100 py-2 px-6 items-center justify-between">
-          <div className="font-semibold text-sm">Model</div>
+          <div className="font-semibold text-sm">{tNav("model")}</div>
           <div className="text-xs">
             <select className="border rounded-md p-2">
               <option value="gpt-3.5">gpt-3.5</option>
@@ -65,7 +70,7 @@ const Navbar: React.FC = () => {
           </div>
         </div>
         <div className="border-b flex border-slate-100 py-2 px-6 items-center justify-between">
-          <div className="font-semibold text-sm">Theme</div>
+          <div className="font-semibold text-sm">{tNav("theme")}</div>
           <div className="text-xs">
             <select className="border rounded-md p-2">
               <option value="auto">auto</option>