+++ components/Welcome/index.tsx
@@ -1,10 +1,12 @@
-import React from "react";
+import * as React from "react";
+import { useTranslation } from "next-i18next";
 import { useOpenAIKey } from "@/hooks";
 
 /**
  * when the OpenAI Key in empty, show the welcome page
  */
 const Welcome: React.FC = () => {
+  const { t } = useTranslation("welcome");
   const [openAIKey] = useOpenAIKey();
 
   if (openAIKey) return null;
@@ -12,16 +14,11 @@ const Welcome: React.FC = () => {
   return (
     <div className="flex h-full justify-center items-center">
       <div className="flex flex-col w-80 md:w-auto">
-        <div className="font-bold text-4xl">Welcome to L-GPT</div>
-        <div className="mt-5 text-lg mb-3">
-          L-GPT is an open-source project that imitates the OpenAI ChatGPT. The
-          theme adopts the style of Feishu.
-        </div>
-        <div className="mb-3 text-gray-500">
-          Please set your OpenAI API Key in the top right of the navbar.
-        </div>
+        <div className="font-bold text-4xl">{t("welcome")}</div>
+        <div className="mt-5 text-lg mb-3">{t("desc")}</div>
+        <div className="mb-3 text-gray-500">{t("set-openai-key")}</div>
         <div className="text-gray-500">
-          If you don't have an OpenAI API Key, you can get one here:{" "}
+          {t("apply-openai-key")}
           <a
             className="text-blue-500 hover:underline"
             href="https://platform.openai.com/account/api-keys"