+++ pages/_app.tsx
@@ -1,13 +1,14 @@
 import type { AppProps } from "next/app";
 import Head from "next/head";
 import { Inter } from "next/font/google";
+import { appWithTranslation } from "next-i18next";
 import { Toaster } from "react-hot-toast";
 
 import "@/styles/globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
 
-export default function App({ Component, pageProps }: AppProps) {
+const App = ({ Component, pageProps }: AppProps) => {
   return (
     <>
       <Head>
@@ -21,7 +22,7 @@ export default function App({ Component, pageProps }: AppProps) {
           content="height=device-height ,width=device-width, initial-scale=1, user-scalable=no"
         />
         <meta name="apple-mobile-web-app-capable" content="yes" />
-        <meta name="apple-mobile-web-app-title" content="L-GPT"></meta>
+        <meta name="apple-mobile-web-app-title" content="L-GPT" />
         <link rel="icon" href="/favicon.svg" />
       </Head>
       <div className={inter.className}>
@@ -30,4 +31,6 @@ export default function App({ Component, pageProps }: AppProps) {
       </div>
     </>
   );
-}
+};
+
+export default appWithTranslation(App);