+++ pages/index.tsx
@@ -1,4 +1,6 @@
-import React from "react";
+import * as React from "react";
+import { GetServerSideProps } from "next";
+import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { Menu, MobileMenu, Navbar, Welcome, ChatSection } from "@/components";
 
 export default function Home() {
@@ -14,3 +16,19 @@ export default function Home() {
     </div>
   );
 }
+
+export const getServerSideProps: GetServerSideProps = async ({ locale }) => {
+  return {
+    props: {
+      ...(await serverSideTranslations(locale ?? "en", [
+        "common",
+        "menu",
+        "chat",
+        "prompt",
+        "nav",
+        "welcome",
+      ])),
+      // Will be passed to the page component as props
+    },
+  };
+};