+++ CHANGE_LOG.md
@@ -4,7 +4,11 @@
 
 > 2023-04-16
 
+### Add
+
+- feat: add en/cn language support
+
 ### Changed
 
-- fixed pnpm-locl.yaml bug
-- fixed @types/math-random bug
+- fixed:pnpm-locl.yaml bug
+- fixed:@types/math-random bug